feat: update sbt to 1.2.0
